rename RealApp to App and App to Main
